+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All Notable changes to yii2-select2-widget will be documented in this file.
 
+## Dev
+
+Assets Update (Xpbl4)
+- Bootstrap Theme
+- Select2 Width
+- Select2 CSS Class
+
 ## 0.1.0
 
 - Initial release (Xpbl4)